functions/requirements.txt
azure-functions==1.12.0
azure-cosmos==4.3.0
requests==2.28.1